+++ R/prediction_mclogit.R
@@ -1,3 +1,3 @@
 #' @rdname prediction
 #' @export
-prediction.mclogit <- prediction.glm
+prediction.mclogit <- prediction.default