+++ R/print.R
@@ -1,84 +1,79 @@
-#' @importFrom utils head
-#' @importFrom stats aggregate
-#' @export
-summary.prediction <- function(object, digits = 4, ...) {
-    # summary method
-    # function also called by `print.prediction()` if object has an 'at' specification
-    
-    # gather metadata
-    f <- object[["fitted"]]
-    fc <- object[["fitted.class"]]
-    vc <- attributes(object)[["vcov"]]
-    
-    # convert 'at_specification' into data frame
-    at <- attributes(object)[["at"]]
-    if (is.null(at)) {
-        objectby <- list(rep(1L, nrow(object)))
-    } else {
-        objectby <- object[ , setdiff(names(at), "index"), drop = FALSE]
-    }
-    
-    # calculate average/modal predictions
-    if (!"fitted.class" %in% names(object) || is.list(fc)) {
-        # numeric outcome
-        ## aggregate average predictions from data
-        out <- aggregate(object[["fitted"]], objectby, FUN = mean, na.rm = TRUE)
-        ## extract calculated variance from object
-        out$SE <- sqrt(vc)
-        
-        # message
-        message(paste0("Data frame with average ", ngettext(nrow(out), "prediction", "predictions"), 
-                " for ", length(f)/nrow(out), " ", 
-                ngettext(nrow(out), "observation", "observations"), ":"))
-    } else {
-        # factor outcome
-        out <- aggregate(object[["fitted.class"]], objectby, FUN = function(set) names(sort(table(set), decreasing = TRUE))[1L])
-        
-        # message
-        message(paste0("Data frame with modal ", ngettext(nrow(out), "prediction", "predictions"), 
-                " (of ", nlevels(factor(fc)), " ", ngettext(nlevels(factor(fc)), "level", "levels"), 
-                ") for ", length(fc), " ", ngettext(length(fc), "observation", "observations"), ": "))
-    }
-    
-    # cleanup output
-    names(out)[!names(out) %in% c("x", "SE")] <- paste0("at(", names(out)[!names(out) %in% c("x", "SE")], ")")
-    names(out)[names(out) == "x"] <- "Prediction"
-    if (is.null(at)) {
-        out <- out[, c("Prediction", "SE"), drop = FALSE]
-    }
-    
-    # print and return
-    print(out, digits = digits, row.names = FALSE, ...)
-    invisible(out)
-}
-
 #' @export
 print.prediction <- function(x, digits = 4, ...) {
     
     # gather metadata
     f <- x[["fitted"]]
     fc <- x[["fitted.class"]]
+    ## at
     at <- attributes(x)[["at"]]
-    vc <- attributes(x)[["vcov"]]
+    at_names <- setdiff(names(attr(x, "at")), "index")
+    
+    ## weights
+    is_weighted <- attr(x, "weighted")
+    if (isTRUE(is_weighted)) {
+        wts <- x[["_weights"]]
+    }
     
     # calculate overall predictions
     ## if no 'at_specification', simply calculate overall average/mode and print
     if (is.null(at)) {
         # object is a single replication with no 'at' specification
-        if (!"fitted.class" %in% names(x) || is.list(fc)) {
-            # numeric outcome
-            m <- sprintf(paste0("%0.", digits, "f"), mean(f, na.rm = TRUE))
-            message(paste0("Data frame with average prediction for ", length(f), " ", ngettext(length(f), "observation", "observations"),
-                           ": ", m, " (se = ", sprintf(paste0("%0.", digits, "f"), sqrt(vc)), ")"))
-        } else {
+        if ("fitted.class" %in% names(x) || is.list(fc)) {
             # factor outcome
             m <- sort(table(x[["fitted.class"]]), decreasing = TRUE)[1L]
-            message(paste0("Data frame with modal prediction (of ", nlevels(factor(fc)), " ", ngettext(nlevels(f), "level", "levels"), 
-                    ") for ", length(fc), " ", ngettext(length(fc), "observation", "observations"), ": ", shQuote(names(m))))
+            message(
+              sprintf("Data frame with %d %s%swith modal prediction (of %d %s):",
+                      length(fc),
+                      ngettext(length(fc), "prediction", "predictions"),
+                      if (!is.null(attr(x, "call"))) sprintf(" from\n %s\n", paste0(deparse(attr(x, "call")), collapse = "\n")) else "",
+                      nlevels(factor(fc)),
+                      ngettext(nlevels(f), "level", "levels"),
+                      shQuote(names(m))
+              )
+            )
+        } else {
+            # numeric outcome
+            message(
+              sprintf("Data frame with %d %s%swith average prediction: %s",
+                      length(f),
+                      ngettext(length(fc), "prediction", "predictions"),
+                      if (!is.null(attr(x, "call"))) sprintf(" from\n %s\n", paste0(deparse(attr(x, "call")), collapse = "\n")) else "",
+                      sprintf(paste0("%0.", digits, "f"), mean(f, na.rm = TRUE))
+              )
+            )
         }
     } else {
         # otherwise, object has an 'at' specification, reflecting multiple requested predictions
-        summary(object = x, ...)
+        
+        # convert 'at_specification' into data frame
+        xby <- x[ , setdiff(names(at), "index"), drop = FALSE]
+        
+        if ("fitted.class" %in% names(x) || is.list(fc)) {
+            # factor outcome
+            out <- aggregate(x[["fitted.class"]], xby, FUN = function(set) names(sort(table(set), decreasing = TRUE))[1L])
+            message(
+              sprintf("Data frame with %d %s%swith modal %s (of %d %s):",
+                      nrow(x),
+                      ngettext(nrow(x), "prediction", "predictions"),
+                      if (!is.null(attr(x, "call"))) sprintf(" from\n %s\n", paste0(deparse(attr(x, "call")), collapse = "\n")) else "",
+                      ngettext(nrow(out), "prediction", "predictions"),
+                      nlevels(factor(fc)),
+                      ngettext(nlevels(fc), "level", "levels")
+              )
+            )
+        } else {
+            # numeric outcome
+            out <- aggregate(x[["fitted"]], xby, FUN = mean, na.rm = TRUE)
+            message(
+              sprintf("Data frame with %d %s%swith average %s:",
+                      nrow(x),
+                      ngettext(nrow(x), "prediction", "predictions"),
+                      if (!is.null(attr(x, "call"))) sprintf(" from\n %s\n", paste0(deparse(attr(x, "call")), collapse = "\n")) else "",
+                      ngettext(nrow(out), "prediction", "predictions")
+              )
+            )
+        }
+        print(out, digits = digits, row.names = FALSE, ...)
     }
     
     # return invisibly